Update config builder
(cherry picked from commit 364473a2bafbfb8c82dd54977d3b240f6095e841)

+++ common/src/main/java/com/sonicether/soundphysics/mixin/SoundEventMixin.java
@@ -14,7 +14,7 @@ public class SoundEventMixin {
         if (!SoundPhysicsMod.CONFIG.enabled.get()) {
             return value;
         }
-        return value * SoundPhysicsMod.CONFIG.soundDistanceAllowance.get().floatValue();
+        return value * SoundPhysicsMod.CONFIG.soundDistanceAllowance.get();
     }
 
 }
